+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.3 [unreleased]
 
+- Added support for materialized views
 - Better handle case when multiple indexes are found for a query
 - Added `--min-cost-savings-pct` option